scrap/requirements.txt
python-dotenv==0.16.0
selenium==3.141.0
urllib3==1.26.4